Split controller into multiple controllers
