--- WalksAndBenches/Services/IStorage.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace WalksAndBenches.Services
-{
-    public interface IStorage
-    {
-    }
-}
