+++ WalksAndBenches/Views/Shared/_Layout.cshtml
@@ -16,8 +16,8 @@
             <ul class="navbar-nav">
                 <li class="nav-item"><a class="nav-link" asp-controller="App" asp-action="Index">Home</a></li>
                 <li class="nav-item"><a class="nav-link" asp-controller="App" asp-action="About">About</a></li>
-                <li class="nav-item"><a class="nav-link" asp-controller="App" asp-action="SubmitWalk">Submit Walk</a></li>
-                <li class="nav-item"><a class="nav-link" asp-controller="App" asp-action="Explore">Explore</a></li>
+                <li class="nav-item"><a class="nav-link" asp-controller="SubmitWalk" asp-action="SubmitWalk">Submit Walk</a></li>
+                <li class="nav-item"><a class="nav-link" asp-controller="Explore" asp-action="Explore">Explore</a></li>
             </ul>
         </nav>
     </header>
